+++ blacksheep/server/bindings/__init__.py
@@ -1,17 +1,16 @@
 """
 This module implements a feature inspired by "Model Binding" in ASP.NET web framework.
-It provides a strategy to have request parameters read an injected into request
+It provides a strategy to have request parameters read and injected into request
 handlers. This feature is also useful to generate OpenAPI Documentation (Swagger)
 automatically.
 
 See:
-    https://docs.microsoft.com/en-us/aspnet/core/mvc/models/model-binding?view=aspnetcore-2.2
+    https://www.neoteroi.dev/blacksheep/binders/
 """
 
 from abc import abstractmethod
-from base64 import urlsafe_b64decode
 from collections.abc import Iterable as IterableAbc
-from datetime import date, datetime
+from functools import partial
 from typing import (
     Any,
     Callable,
@@ -27,16 +26,14 @@ from typing import (
     TypeVar,
     Union,
 )
-from urllib.parse import unquote
-from uuid import UUID
 
-from dateutil.parser import parse as dateutil_parser
 from guardpost import Identity
 from rodi import CannotResolveTypeException, ContainerProtocol
 
 from blacksheep import Request
 from blacksheep.contents import FormPart
 from blacksheep.exceptions import BadRequest
+from blacksheep.server.bindings.converters import converters
 from blacksheep.server.websocket import WebSocket
 from blacksheep.url import URL
 
@@ -146,9 +143,7 @@ class FromHeader(BoundValue[T]):
 
 
 class FromQuery(BoundValue[T]):
-    """
-    A parameter obtained from URL query parameters.
-    """
+    """A parameter obtained from URL query parameters."""
 
 
 class FromCookie(BoundValue[T]):
@@ -293,8 +288,15 @@ class Binder(metaclass=BinderMeta):  # type: ignore
     def is_generic_iterable_annotation(self, param_type):
         return hasattr(param_type, "__origin__") and (
             param_type.__origin__ in {list, tuple, set}
-            or issubclass(param_type.__origin__, IterableAbc)
+            or self._issubclass(param_type.__origin__, IterableAbc)
         )
+
+    @staticmethod
+    def _issubclass(clstype, class_or_tuple) -> bool:
+        try:
+            return issubclass(clstype, class_or_tuple)
+        except TypeError:
+            return False
 
     def generic_iterable_annotation_item_type(self, param_type):
         try:
@@ -406,7 +408,11 @@ def _try_get_type_name(expected_type) -> str:
 def get_default_class_converter(expected_type):
     def converter(data):
         try:
-            return expected_type(**data)
+            if isinstance(data, dict):
+                return expected_type(**data)
+            else:
+                # list, simple type
+                return expected_type(data)
         except TypeError as type_error:
             raise BadRequest(
                 "invalid parameter in request payload, "
@@ -436,29 +442,9 @@ class BodyBinder(Binder):
         self.converter = converter
 
     def _get_default_converter_single(self, expected_type):
-        # this method converts an item that was already parsed to a supported type,
-        # for example in JSON can be int, float, bool
-        if expected_type in {str, int, float, bool} or str(expected_type) == "~T":
-            return lambda value: value
-
-        if expected_type is date:
-            return lambda value: dateutil_parser(value).date() if value else None
-
-        if expected_type is datetime:
-            return lambda value: dateutil_parser(value) if value else None
-
-        if expected_type is bytes:
-            # note: the code is optimized for strings here, not bytes
-            # since most of times the user will want to handle strings
-            return lambda value: (
-                urlsafe_b64decode(value.encode("utf8")).decode("utf8")
-                if value
-                else None
-            )
-
-        if expected_type is UUID:
-            return lambda value: UUID(value)
-
+        for converter in converters:
+            if converter.can_convert(expected_type):
+                return partial(converter.convert, expected_type=expected_type)
         return get_default_class_converter(expected_type)
 
     def _get_default_converter_for_iterable(self, expected_type):
@@ -601,29 +587,11 @@ class BytesBinder(Binder):
         return await request.read()
 
 
-def _default_bool_converter(value: str) -> bool:
-    if value in {"1", "true"}:
-        return True
-
-    if value in {"0", "false"}:
-        return False
-
-    # bad request: expected a bool value, but
-    # got something different that is not handled
-    raise BadRequest(f"Expected a bool value for a parameter, but got {value}.")
-
-
-def _default_bool_list_converter(values: Sequence[str]):
-    return _default_bool_converter(values[0].lower()) if values else None
-
-
 class SyncBinder(Binder):
     """
     Base binder class for values that can be read synchronously from requests
     with complete headers. Like route, query string and header parameters.
     """
-
-    _simple_types = {int, float, bool}
 
     def __init__(
         self,
@@ -631,82 +599,45 @@ class SyncBinder(Binder):
         name: str = "",
         implicit: bool = False,
         required: bool = False,
-        converter: Optional[Callable] = None,
+        converter: Optional[Callable[[Sequence[str]], Any]] = None,
     ):
         super().__init__(
             expected_type,
             name=name,
             implicit=implicit,
             required=required,
-            converter=converter or self._get_default_converter(expected_type),
+            converter=converter or self._get_converter(expected_type),
         )
 
-    def _get_default_converter_single(self, expected_type):
-        if expected_type is str or str(expected_type) == "~T":
-            return lambda value: unquote(value) if value else None
-
-        if expected_type is bool:
-            return _default_bool_converter
-
-        if expected_type is bytes:
-            # note: the code is optimized for strings here, not bytes
-            # since most of times the user will want to handle strings
-            return lambda value: value.encode("utf8") if value else None
-
-        if expected_type in self._simple_types:
-            return lambda value: expected_type(value) if value else None
-
-        if expected_type is UUID:
-            return lambda value: UUID(value)
-
-        if expected_type is datetime:
-            return lambda value: dateutil_parser(unquote(value)) if value else None
-
-        if expected_type is date:
-            return lambda value: (
-                dateutil_parser(unquote(value)).date() if value else None
-            )
-
-        raise MissingConverterError(expected_type, self.__class__)
-
-    def _get_default_converter_for_iterable(self, expected_type):
-        generic_type = self.get_type_for_generic_iterable(expected_type)
-        item_type = self.generic_iterable_annotation_item_type(expected_type)
-        item_converter = self._get_default_converter_single(item_type)
-        return lambda values: generic_type(item_converter(value) for value in values)
-
-    def _get_default_converter(self, expected_type):
-        if expected_type is str or str(expected_type) == "~T":
-            return lambda value: unquote(value[0]) if value else None
-
-        if expected_type is bool:
-            return _default_bool_list_converter
-
-        if expected_type is bytes:
-            return lambda value: value[0].encode("utf8") if value else None
-
-        if expected_type in self._simple_types:
-            return lambda value: expected_type(value[0]) if value else None
-
-        if expected_type is UUID:
-            return lambda value: UUID(value[0]) if value else None
-
+    def _get_converter(self, expected_type) -> Callable[[Sequence[str]], Any]:
         if self.is_generic_iterable_annotation(expected_type) or expected_type in {
             list,
             set,
             tuple,
         }:
-            return self._get_default_converter_for_iterable(expected_type)
+            return self._get_converter_for_iterable(expected_type)
 
-        if expected_type is datetime:
-            return lambda value: dateutil_parser(unquote(value[0])) if value else None
-
-        if expected_type is date:
-            return lambda value: (
-                dateutil_parser(unquote(value[0])).date() if value else None
-            )
+        for converter in converters:
+            if converter.can_convert(expected_type):
+                return lambda values: converter.convert(
+                    values[0] if values else None, expected_type
+                )
 
         raise MissingConverterError(expected_type, self.__class__)
+
+    def _get_converter_single(self, expected_type):
+        for converter in converters:
+            if converter.can_convert(expected_type):
+                return partial(converter.convert, expected_type=expected_type)
+        raise MissingConverterError(expected_type, self.__class__)
+
+    def _get_converter_for_iterable(
+        self, expected_type
+    ) -> Callable[[Sequence[str]], Any]:
+        generic_type = self.get_type_for_generic_iterable(expected_type)
+        item_type = self.generic_iterable_annotation_item_type(expected_type)
+        item_converter = self._get_converter_single(item_type)
+        return lambda values: generic_type(item_converter(value) for value in values)
 
     @abstractmethod
     def get_raw_value(self, request: Request) -> Sequence[str]:
@@ -723,12 +654,10 @@ class SyncBinder(Binder):
         return value in self._empty_iterables
 
     async def get_value(self, request: Request) -> Optional[Any]:
-        # TODO: support get_raw_value returning None, to not instantiate lists
-        # when a parameter is not present
         raw_value = self.get_raw_value(request)
         try:
             value = self.converter(raw_value)
-        except (ValueError, BadRequest) as converter_error:
+        except ValueError as converter_error:
             raise BadRequest(
                 f"Invalid value {raw_value} for parameter `{self.parameter_name}`; "
                 f"expected a valid {self.expected_type.__name__}."
@@ -791,7 +720,7 @@ class RouteBinder(SyncBinder):
     def __init__(
         self,
         expected_type: T = str,
-        name: str = None,
+        name: Optional[str] = None,
         implicit: bool = False,
         required: bool = True,
         converter: Optional[Callable] = None,